+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Deprecated `craft\commerce\models\ProductType::lineItemFormat`.
 
 ### Fixed
+- Fixed a bug where the Variant Field could be added to a Products’ field layout twice.
 - Fixed a bug where PDFs were being rendered incorrectly. ([#2599](https://github.com/craftcms/commerce/issues/2599))
 - Fixed a PHP that that could occur when instantiating an Order object. ([#2602](https://github.com/craftcms/commerce/issues/2602))